+++ blueprint/iam-base/registrations.ts
@@ -121,12 +121,12 @@ const environmentConfig = configInjector.chain({
 
 //! defines the runtime dependencies for the application
 const runtimeDependencies = environmentConfig.chain({
-  MikroORM: {
+  Orm: {
     lifetime: Lifetime.Singleton,
     type: MikroORM,
     factory: () => new MikroORM(mikroOrmOptionsConfig)
   },
-  OpenTelemetryCollector: {
+  OtelCollector: {
     lifetime: Lifetime.Singleton,
     type: OpenTelemetryCollector<Metrics>,
     factory: ({ OTEL_SERVICE_NAME, OTEL_LEVEL }) =>
@@ -140,10 +140,10 @@ const runtimeDependencies = environmentConfig.chain({
     lifetime: Lifetime.Scoped,
     type: EntityManager,
     factory: (
-      { MikroORM },
+      { Orm },
       context: { entityManagerOptions?: ForkOptions; tenantId?: string }
     ) => {
-      const em = MikroORM.em.fork(context.entityManagerOptions);
+      const em = Orm.em.fork(context.entityManagerOptions);
       if (context.tenantId) {
         em.setFilterParams('tenant', { tenantId: context.tenantId });
       }
@@ -162,12 +162,12 @@ const serviceDependencies = runtimeDependencies.chain({
       OrganizationMapperTypes,
       OrganizationDtoTypes
     >,
-    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
+    factory: ({ EntityManager, OtelCollector }, context, resolve) =>
       new BaseOrganizationService(
         context?.entityManagerOptions
           ? resolve?.('EntityManager', context)
           : EntityManager,
-        OpenTelemetryCollector,
+        OtelCollector,
         schemaValidator,
         {
           OrganizationMapper,
@@ -183,13 +183,13 @@ const serviceDependencies = runtimeDependencies.chain({
       PermissionMapperTypes,
       PermissionDtoTypes
     >,
-    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
+    factory: ({ EntityManager, OtelCollector }, context, resolve) =>
       new BasePermissionService(
         context.entityManagerOptions
           ? resolve('EntityManager', context)
           : EntityManager,
         () => resolve('RoleService', context),
-        OpenTelemetryCollector,
+        OtelCollector,
         schemaValidator,
         {
           PermissionMapper,
@@ -202,12 +202,12 @@ const serviceDependencies = runtimeDependencies.chain({
   RoleService: {
     lifetime: Lifetime.Scoped,
     type: BaseRoleService<SchemaValidator, RoleMapperTypes, RoleDtoTypes>,
-    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
+    factory: ({ EntityManager, OtelCollector }, context, resolve) =>
       new BaseRoleService(
         context.entityManagerOptions
           ? resolve('EntityManager', context)
           : EntityManager,
-        OpenTelemetryCollector,
+        OtelCollector,
         schemaValidator,
         {
           RoleMapper,
@@ -224,12 +224,12 @@ const serviceDependencies = runtimeDependencies.chain({
       UserMapperTypes,
       UserDtoTypes
     >,
-    factory: ({ EntityManager, OpenTelemetryCollector }, context, resolve) =>
+    factory: ({ EntityManager, OtelCollector }, context, resolve) =>
       new BaseUserService(
         EntityManager,
         () => resolve('RoleService', context),
         () => resolve('OrganizationService', context),
-        OpenTelemetryCollector,
+        OtelCollector,
         schemaValidator,
         {
           UserMapper,
@@ -241,16 +241,16 @@ const serviceDependencies = runtimeDependencies.chain({
   ComplianceDataService: {
     lifetime: Lifetime.Singleton,
     type: ComplianceDataService,
-    factory: ({ MikroORM, OpenTelemetryCollector }) =>
-      new ComplianceDataService(MikroORM, OpenTelemetryCollector, {
+    factory: ({ Orm, OtelCollector }) =>
+      new ComplianceDataService(Orm, OtelCollector, {
         User: 'id'
       })
   },
   RetentionService: {
     lifetime: Lifetime.Singleton,
     type: RetentionService,
-    factory: ({ MikroORM, OpenTelemetryCollector }) =>
-      new RetentionService(MikroORM, OpenTelemetryCollector)
+    factory: ({ Orm, OtelCollector }) =>
+      new RetentionService(Orm, OtelCollector)
   }
 });