+++ blueprint/iam-better-auth/api/controllers/user.controller.ts
@@ -7,7 +7,7 @@ import {
 } from '@forklaunch/blueprint-core';
 import { ci, tokens } from '../../bootstrapper';
 
-const openTelemetryCollector = ci.resolve(tokens.OpenTelemetryCollector);
+const openTelemetryCollector = ci.resolve(tokens.OtelCollector);
 const HMAC_SECRET_KEY = ci.resolve(tokens.HMAC_SECRET_KEY);
 const serviceFactory = ci.scopedResolver(tokens.SurfacingService);