+++ blueprint/implementations/worker/redis/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @forklaunch/implementation-worker-redis
 
+## 1.0.22
+
+### Patch Changes
+
+- update internal packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.20
+
+## 1.0.21
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.19
+
+## 1.0.20
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.18
+
+## 1.0.19
+
+### Patch Changes
+
+- update package versions
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.17
+
+## 1.0.18
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.16
+
+## 1.0.17
+
+### Patch Changes
+
+- bump packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.15
+
+## 1.0.16
+
+### Patch Changes
+
+- update packages
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.14
+
 ## 1.0.15
 
 ### Patch Changes